Added study log and a few other things
added study log, which saves information about the previous 50 study sessions to the user's device, and displays them on the study log page. also made it possible to return to the screen you came from on the about this site page, instead of having to go to main menu. same thing works for the study log page